Update logic when the tandem tool changes RMG tolerances
Previously, the tandem tool allowed the user to specify the RMG tolerance in the arc.yml file. However, this only changed toleranceMoveToCore. As described in page 51 from the RMG-Py Documentation Handbook, when not pruning it is recommended to set both toleranceMoveToCore and toleranceInterruptSimulation to the same value.
When pruning, the algorithm only prunes if integration is complete so it is recommended to have toleranceInterruptSimulation be some large number such as  1e8 or 5e8 to effectively prevent the simulation from being interrupted. 
This logic is now included in the tandem tool.