Fix isomorphism checking in databaseTest.
Previously, if there was an non-isomorphism present in the 
labeled atoms, then the mapping would be invalid, but this 
is not (currently) checked by the is_isomorphic method,
so two non-isomorphic molecules would be found isomorphic,
and would fail the test.

Now we ensure that the mapping is valid, before passing it 
to is_isomorphic. This should work whatever the outcome of 
the discussion around PR #1894