Update write_to_database methods
The string representations of level of theory objects used in the
database might differ from those used in AE and BAC classes. However,
write_to_database has to edit the database file as a text file instead
of treating it as a Python module, so make sure to match the correct
string representation of the level of theory in the database.

UPDATE: upon rebasing (3 years later), it looks to me (rwest) like this commit is
redundant, so by the time I resolved the conflicts, as far as I can
tell, there is nothing left. But I'm leaving the message here.
The original commit was eabdee035475041d7f171ea18a113eb8381b639a