Update anacondaDeveloper.rst
Combine Ubuntu installation commands for gcc, g++, git and make to one line;
add equivalent for installation instruction for Fedora since this is a major distro on the Red Hat side (vs. Debian side of the Linux world with Ubuntu) and g++ is not required as a separate package.
This should add clarity for newer users.