When manipulating a graph, reset the connectivity values.
If you have added or removed a vertex or edge on a graph,
then the connectivity values are most likely invalid and
should be reset.
Special case is adding a vertex to a graph, only the vertex
needs resetting (although the consequences are probably the same).

This should prevent weird bugs in isomorphism. The connectivity
values are used in isomorphism checks, and if they are wrong
then you get incorrect results.