Make the reaction library writing test more general and test a surface example.
This commit generalizes the test and makes it also test the write-read loop for
a surface mechanism (which was copied from RMG-database/input/kinetics/libraries/Surface/Example.

The test however is not good enough, as it doesn't actually enforce the written library
is the same as the input. For example, the `metal` attribute in the original is not
duplicated in the copy, and yet the test still passes. 
This may indicate a problem with a __repr__ method somewhere as well as the saving.