Fix apparent typo in evaluating net charge of Group templates after reaction.
Based on the context of the code, it's pretty clear this is a typo,
and we should be summing over the variable that is being updated within
the loop (not some other random thing that's in memory from earlier).

But unfortunately, fix ing it seems to reveal an error somewhere else.

Now in the <ReactionFamily "1,2_Insertion_CO"> 
when trying to generate the template for the reverse 
[<Entry index=0 label="COS">, <Entry index=1 label="RR'">]
it ends up with the product
1 *1 C                                     u0 p0 c+0.5 {2,[B,D]} {3,S} {4,S}
2 *4 [S,O]                                 u0 p2 c+0.5 {1,[B,D]}
3 *2 [Cb,Sis,O,Cd,Cs,N,S,Ct,Cdc,H,Csc,Sid] u0 {1,S}
4 *3 H                                     u0 {1,S}
which has a net charge.
(I'm not sure where those +0.5 charges come from)