CHANGE_BOND with vdw bonds
Revised the change_bond procedure to in family `_apply()` to make a vdW bond, then change it.

In order for the reverse template to work, we need the creation of a single bond 
from a vdW bond to be "increment" from 0, rather than "create".
Thus the CHANGE_BOND needs to sometimes be able to work when there isn't a bond.