When perturbing the TS energy perturb the Ea of the kinetics as well
If k(E) is calculated via Laplace Transform rather than RRKM it's very sensitive to the kinetics
and can have only limited sensitivity to the transition state energy. In this case it makes sense to
perturb the Ea by the same about as the transition state energy (as that is approximately the effect
the TS pertubation should have on the high pressure limit kinetics).