Add data_count attribute to solute group Entry
A new attribute `data_count` is added to `Entry` class. `data_count` is
used to store the number of data used to fit the group values.

`data_count` attribute is used for solute groups in the solvation group
additivity method. It is saved as `DataCountGAV` class, which stores the
number of data used to fit each solute parameter group value.

An unittest has been added to test that solute groups can be loaded with
correct solute parameter values and data counts