Add data_count attribute to solvent Entry
A new class `DataCountSolvent` is created to store the number of
solvation free energy and enthalpy data used to fit solvent parameters.
It also stores the solvation free energy and enthalpy error associated
with the solvent parameters.

This class is added as a `data_count` for `SolventLibrary` that stores
each solvent entry.

The relevant unittest has been added. Some of the existing unittest
values have been changed as the solvent parameter data got updated in
RMG-database.