Optionally `generate_resonance_structures` when loading species dicts
Resonance structure generation can slow down loading of species dictionaries.
Usually this is not significant, but it can be if you are loading a very large
species dictionary from a model's edge. This makes it optional 
(with defaulting to True), similar to how it is for get_species method for database.