check multiplicity of products if the family template reactants restrict multiplicity
If applying a family recipe in reverse and the template reactant groups restrict multiplicity,
we need to make sure that a reverse product matches the multiplicity-constrained template reactant
because the forward template products do not enforce the multiplicity restriction. This commit
adds a check for this in the apply_recipe method and returns `None` (no reaction) if
none of the products generated match the multiplicity-constrained template reactant.