Test documentation on pull requests.
We only publish the documentation from a push event to the official master,
but we now build and then test the documentation from pull requests.

The -W flag turns warnings into errors, and the --keep-going will make sure
it doesn't exit with non-zero exit code until the end (so you get all the errors
not just the first).