Don't generate the debug messages if we're not going to log them.
Should speed it up a bit. 
This might not be called often enough for this optimization to matter,
but I'm not sure. At first I just wanted to switch to %s formatting
because it uses lazy evaluation of the sting formatting, but that still
leaves the expensive to_adjacency_list() calls and the only way to avoid
those is wrap the whole thing an if block.