Update QChem parser
Fixes a bug in which 0 could be returned as an energy if a=0 and b is another value. Updates the doc string to accurately reflect the method. Note that this parses the energy slightly differently than was implemented in Grambow et al. Here, the energy from the geometry optimization is prioritized while Grambow et al. parsed the last energy in the file, often from a frequency calculation

https://github.com/cgrambow/ard_gsm/blob/82826011b0edf7122e16063094b04ecd16bf1cf5/ard_gsm/qchem.py#L82-L86

https://github.com/cgrambow/ard_gsm/blob/82826011b0edf7122e16063094b04ecd16bf1cf5/ard_gsm/qchem.py#L149-L154

-1 is used when indexing to improve code readability