QChemLog: fix energy parsing logic
Previous logic for extracting the energy would fail because by setting `a = b = 0`, `e_elect` was never `None`. Thus, the `LogError` was never raised, even if no matches were found. Instead, 0 was returned as the molecular energy, which impacted downstream calculations. This is resolved by initializing the variables to None.

As a style choice, it is easier to read "take the last value after splitting" rather than "take the 3rd or 8th value after splitting"