More diagnostic error message for debugging restart from seed with filters.
I was getting a weird error in the functional test:
ERROR: Test that the RMG restart job with filters included completed without problems
  File "/Users/rwest/Code/RMG-Py/rmgpy/rmg/mainTest.py", line 210, in test_restart_with_filters
so made the error message more informative.
The problem went away when I updated the RMG-database to the latest, but we might as
well keep the more detailed error message.