add reaction systems from input to reaction model
When species are removed from the edge, rmg removes them from the reaction systems attached to the CoreEdgeReactionModel.
The reaction systems are added to the main RMG object in rmg.input (rmg.reaction_systems).
This commit adds the reaction systems to the reaction model so that edge species can be removed using the reaction model's `remove_species_from_edge` method.
Otherwise, rmg.reaction_model.reaction_systems is an empty list, so we don't actually remove the edge species from the reaction systems.