bugfix - removing specs from pdep network
A species may be in a net reaction and not a path reaction if a species in the network was previously removed from the edge.  Therefore, we need to check if the species is in a net reaction even if the species is not in any path reactions. The species should also be removed from list of explored since the species will be appended to network isomer list with update_configurations.  Also, `remove_empty_pdep_networks` method needs to be called after `remove_species_from_edge` in `add_species_to_core` if the spec is forbidden structure.