Correct way to handle adding new reactions to core
Check if the rxn we want to add to rmg core is already in rmg edge. If the rxn is already in rmg edge, we should pass that rxn from edge phase to core phase, instead of creating a new rxn object and directly put in core phase. If the rxn we want to add to rmg core is not in rmg edge, then we should put the newly created rxn to both core phase and edge phase, instead of creating one new reaction for the core phase and one new reaction for the edge phase.