Build edge phase first
We have to build the edge phase first, because the efficiencies dictionary in any kinetics that has efficiencies gets re-written every time we build a phase. If we build the edge phase later, the nameefficiencies will contain edge species, and the efficienies dictionary will have edge species index. This ultimately causes the core simulation to crash, if any of the efficiency species are still in edge.