revised arrbm `get_activation_energy` method
We previously did not account for E0 < 0.  Here, for E0 < 0, we set Ea to E0 for exothermic rxns and dHrxn for endothermic.  We also incorporate aspects of Reaction `fix_barrier_height` method to set Ea to 0 for exothermic rxns and Ea < 0, and set Ea to dHrxn if Ea < dHrxn for endothermic rxns.