Prevent use of certain methods with product channels
CSE-G as defined cannot be used with product channels

SLS in theory should be able to handle product channels, but it can give some crazy results currently.
My suspicion is that this isn't due to method limitations and is in fact a bug in the sls code that
only triggers when product channels are present. However, at present I have been unable to find
the hypothetical bug. In practice, though, this is not usually an important limitation, in fact MESS uses
the CSE-G method.

fix input error test