BugFix: avoid changing atom order in analyze_molecule
Previously, generate_resonance_structures is fixed to allow atom order being preserved. However, there are still a few cases atom order are still changed. According to an initial investigation, most of them are ring molecules, and the change takes place in the `get_aromatic_rings`. Changes are made to correct the issue.