Set Keq=inf for irreversible reactions.
Third body and pressure dependent reactions with collision efficiencies
need to have their rate constants recalculated during the solver step because
of the efficiencies, which means their reverse rates are re-evaluated
using the Keq. If the Keq was left as zero (as the case prior to this commit)
because a reaction was irreversible, then it would crash with a zero division
error.

This fix allows irreversible, pressure-dependent reactions, with collision efficiencies.

(Not a common scenario - why would a PDep reaction ever be irreversible?
but some published kinetic models have them, and so they have crept in
to user's seed mechanisms).