Make 'test' a phony target in Makefile
Now that there's a folder called 'test' the target with
the same name doesn't work properly in make, i.e.
if you do `make test` it just checks that the 'test'
folder is up to date. 

My making it a "phony" target it doesn't do this, and it 
once again works as intended.