run CI on PRs as well as push
this will allow the status of the PR to be reported to RMG-Py, since the push action only reports the status to the base repo. this will cause the CI to simultaneously run on both the fork and RMG-Py, but that will not affect the total number of allowed runners.