Remove redundant explanation about linking python to python-jl
This explanation was added in commit 5d208269ce5dabc73d3152170ac75f1dfbce4bd1
but I think it refers to the line
   ln -sfn $(which python-jl) $(which python) 
that was added in b6910a76fdef684995a886384f50bbdd2eace901 (2 minutes later)
to the continuous integrations workflow, but not actually added to the
user instructions. i.e. the linking step (that this explanation refers to) 
is not in the instructions. 

So I suggest we remove it.