CI: Don't create symbolic link from python-jl to python
We used to force a symlink of python-jl to where python used to be,
so that when you ask for python you get python-jl.
This would work if the python-jl itself called to python3, (which is 
what our hacked version of it did). But the conda-forge version of 
python-jl calls python, and if python points to python-jl, then
you get an infinite loop of calling python-jl.

That's my hunch. This will test the theory, and see what happens.
Maybe we'll have to explicitly call python-jl in a few places we
used to be lazy and called python. But it seems to work and makes
us compatible with a standard python-jl.
Also, "explicit is better than implicit!"