Print statements about using the Julia system image.
I wanted to use logging, but this is imported before the logging
module is imported and set up.

Strangely, using the system image I made like this in julia:
 using PackageCompiler
 using ReactionMechanismSimulator
 create_sysimage(["ReactionMechanismSimulator"]; sysimage_path="rms.so")
my time to launch is no faster than not having it.
Here are a couple of runs of
python-jl rmg.py test/regression/superminimal/input.py
289.01s user 20.75s system 109% cpu 4:43.95 total
305.97s user 22.11s system 106% cpu 5:09.14 total
compared to without the special system image:
291.06s user 22.26s system 107% cpu 4:52.13 total
in both cases, actually running RMG was about 1 minute.