CI: Test the documentation properly.
The "Test documentation" step wasn't rebuilding the documentation because
it was already built, so wasn't checking anything.
So I replaced the "Make documentation" step with the commands that test it,
for when you're not doing a push to main.
(when you're doing a push to main it just does its best and publishes)