CI docs: separate job step for compiling pyrms
This takes a while and currently looks like it's part of the
making of documentation, but is in fact part of installing pyrms

It'd be nice if this weren't needed.
I *hope* this step just moves the time spent compiling this stuff
and does not duplicate the effort (in which case the step should 
definitely be removed). Maybe it should be removed anyway.