'make documentation' now works on MacOS with new python-jl
Because building the documentation involves importing the classes
and importing from reactors.py involves importing from pyrms and diffeqpy
and because those involve loading julia stuff which involves JIT compiling,
the whole thing crashes if your python version (called by `sphinx-build`) is
statically linked or something. So you have to to run sphinx-build using
python-jl instead of python. 

Which is what this does.