Fix typo in environment.yml
This is a combination of 12 commits, as we tried everything under the sun to debug the CI.
But they all cancelled out when merged (apart from fixing a typo in a comment)

- This is the 1st commit message:
CI: trying to force ubuntu-20.04 instead of ubuntu-latest (temporary)
While we figure out what's happening, let's try an older ubuntu.
- This is commit message #2:
fixup! CI: trying to force ubuntu-20.04 instead of ubuntu-latest (temporary)
- This is the commit message #3:
Switch back to ubuntu-latest
But leave a 20.04 in the matrix build
- This is the commit message #4:
Trying libstdcxx-ng < 13 in Conda environment.
Trying to debug. If this works it should be put in docker file too.
Or, better, the real cause found and fixed.
- This is the commit message #5:
fix typo in env, undo ubuntu os changes, set gcc version in CI to 6
this system object of this version is being provided by gcc, and the
runners no longer come prepackaged with it since its old (?)
- This is the commit message #6:
it wasn't the gcc version
- This is the commit message #7:
it was the julia version, 1.9.0 is brokey
as reported at https://github.com/conda-forge/julia-feedstock/issues/253
the latest release of julia (1.9.0) is brokey, don't use it (put !=1.9.0 in the environment file)
- This is the commit message #8:
make mac and ubuntu use the same cxx library
- This is the commit message #9:
but what if it was rdkit all along?
- This is the commit message #10:
Revert "but what if it was rdkit all along?" because it wasn't
- This is the commit message #11:
Revert "make mac and ubuntu use the same cxx library" cos it didn't work
- This is the commit message #12:
Revert "it was the julia version, 1.9.0 is brokey" but it wasn't