CI: use an env variable to track if running the reference case job
This changes it so that only the ubuntu runner on the official main branch
is the reference case. It also stores that result in an environment variable
so we only need to evaluate the logic once.

For regression testing we need to determine if we're the reference case
(env.REFERENCE_JOB=='true') or the comparison case (env.REFERENCE_JOB=='false')

Because it becomes a string, 'false' is then read as a
non-empty string and evaluates as true.
So we have to do string comparison not negation.
i.e. instead of ${{ !( env.REFERENCE_JOB ) }}
     we must do ${{ env.REFERENCE_JOB == 'false' }}