CI: fix regression testing
The way exit values are returned from shell commands
is not compatible with the way this script had been refactored.
To save the exit code as a variable you'd need to use $? but then
the evaluations etc get more complicated.

What had been implemented was evaluating as a failure
when it passed and a pass when it failed, and was hard to fix.
This may look verbose but is simpler and should work.

It would be nice if all this logic were perhaps in Python,
instead of bash.