CI: don't run regression comparison for RMS_liquidSurface_ch4o2cat
Pull request #2470 shows the difficulty in getting this to work.
It'll be easier once #2335 is merged.
But in the mean time, let's just not try.
We'll still run the test case, so we know it doesn't crash, but we won't try
comparing the output for now, because trying to fix it is holding back our testing.