Changing 'or or' to 'in set' checks.
If you happen to pick the first of the list then the
or or method was faster. But if you pick
the last of the list (or not in the list) then the
in tuple method is faster.

Also changed some lists to sets for consistency,
and because they're a touch faster.

See pylint https://pylint.pycqa.org/en/latest/user_guide/messages/refactor/consider-using-in.html

(Change requested in code review)

Also switched a format to an fstring.