Make check explicitly 'not None'.
The only way r would be 'None' is if a symmetrical reaction (reactants == products)
was found in a library, and hit this check in check_for_existing_reaction that's called by 
make_new_reaction. But that suggests some other problem. So I doubt this test is often needed.
But Matt added it for something.

I can't see any way it would be False, so now we check it's not None.
(as requested in code review)