More detailed regression test results in PR annotation.
The regression-comparison step now captures its STDOUT
(and STDERR) outputs and they are used to create the
annotation on the pull request.

I tried the mathiasvr/command-output@v2.0.0
workflow,  but it doesn't let us use a bash
login shell, and maybe has other vulnerabilities.
So I reverted to bash and trying to just
redirect outputs to files.