Futzing with annotation workflow.
The continue-on-error:true makes it think that step succeeded
(though it didn't) so the "do on failure"
doesn't run, and then the next step fails.

Now we let it fail, and at least report in json
the information that may let us try something else.
One idea is to cange the javascript around
line 32 that gets the matchArtifact
and if there isn't one then return there
and behave differently.