When loading all reaction libraries, only process "reactions.py" files.
I had a reaction library folder with another python file in it
that had been used to renumber the reactions
at some point. RMG was trying to load this
as if it were a reactions.py, and failing the database tests.
Since named reaction libraries are forced to have
the file be called "reactions.py", I think it is OK
that the "load all libraries" method makes the
same assumption.