Update create_augmented_layers
Previously, rdkitmol is created after removing hydrogens from the rmg molecule. This is inconsistent with the assumption of the updated version of to_rdkit_mol, where heavy atoms are assumed to have no implicit hydrogens. This commit moves conversion to rdkitmol and generating aux info prior to the hydrogen removal. It fixes the inconsistency and is possibly more robust, as the complete molecule's definition is used to generate the aux info.