Update the ref inchi used in translatorTest for ch2o2
In some of the early versions of RDKit, a different (but valid) inchi will be generated by RDKit comparing to Openbabel. In a version between 2020.03.3 to 2022.09.1, RDKit changes (potentially improve ) the inchi handling, and now both of them return the same inchi, which is the one we used for unit tests in this commit.