Tidy up reaction.py, removing fix_reaction
This function is only used in one place for one very specific thing
so I inlined it. It also had a vague name.

I then commented it out, because I'm not clear why it's needed.
If it is needed, we can restore it (and put a comment as to why)