Fix to thermoTest TestCyclicThermo::test_remove_group
The test depends on the radical groups tree in the testing 
database being a certain way, which it no longer is.
We used to test it on a different dummy database, which we have
just removed to reduce redundancy, but rather than change the database
to be that way (which messes up other tests) I am modifying the test 
to first put the database (in memory) the way it expects it to be.

The function being tested here (remove_group) is never used in RMG.
It was written in 89901fa97672fcf49bfda16ed17e7e3531bbc926 
by nyee who wrote, in 2016, "Currently, I am using this to remove CS groups as they 
are not detected properly". So we're fixing a broken test for a working
function that is never used.