Better debugging messages for database errors.
When there are issues in a database file, such as nan values here:
  uncertainty=RateUncertainty(mu=nan, var=nan,
which may be due to a bug in the tree fitting, it was very hard to 
find the cause of the error because the stack trace wouldn't tell you
where the error was, and the file handle had been read by the time you
caught the exception.
With this change, we read the file to a variable, then try executing it,
and catch and report any exceptions properly.