Conversion to RDKitMol can use 0 for surface sites X
You can use a * to represent a surface site in a SMILES string when
reading it via RDKit, and RDKit turns this into a dummy atom
with atomic number 0.

By doing the reverse (telling RDKit that our surface sites have
atomic number 0) we can use RDKit to *generate* SMILES strings
in the same format, enabling a round trip.

But for other things like InChIs it seems more robust to use
an atom like Platinum (78). This allows both, with default 0,
but 78 used in InChI conversion.