New unit tests for SMILES conversions of Adsorbates.
Unfortunately going from a molecule TO a smiles uses
OpenBabel if you have Nitrogen in the molecule, which 
then uses [Pt] in place of *.
But you can still READ smiles with * and N in. 
That means you don't get a round trip. 
In [9]: Molecule(smiles='CNC*').to_smiles()
Out[9]: 'CNC[Pt]'

Still, this is better than it was. (I think).