Fixed Fragment test failure
The `Fragment` class is now a subclass of
`Molecule`. As such, some if statements for
checking atom charges no longer work for Fragments
since CuttingLabels don't have defined charges.

These if statements were modified such that charge
checks for `Fragments` ignore CuttingLabels, and
will skip charge checks for `Molecules` if the
species is an instance `Fragments`.