Estimating thermo of an adsorbate Species now re-orders Species.molecule.
We do this in the gas phase, so now also adsorbates:
If you have a Species with multiple Molecules (resonance structures)
then the list of Molecules gets re-ordered when you estimate the thermo
so that the most stable (lowest H298) goes first.
This will then be used for drawing pictures, printing SMILES strings, etc.
which might mean it's no the one that you started with.
This may surprise some people in some situations, but I think is the 
way the gas phase works, and it has some benefits, especially for 
machine-generated species.

One thing we don't yet do in adsorbates is prioritize resonance structures coming
from thermo libraries, over those that are merely estimated.
Not sure of the consequences of this.

Other changes in this commit are mostly for clarity.