In Reaction.to_cantera, don't make kinetics twice.
The method calls the appropriate set_cantera_kinetics() method
at the end, so we don't want to make them before then,
because it's wasteful to do it twice. We just need to create
a cantera kinetics object of the correct type.
This was already done for several types of kinetics,
but not all.