Update H_Abstraction family in test database
We need at least one hand-built kinetics tree with a complicated enough
structure to generate all of the different kinetics source case
scenarios: falling up the tree to where
there is data, falling up the tree on multiple subtrees and averaging
the resulting templates, using a template that only has data averaged
from other nodes, and using a node with data matched to that specific
template. It's hard to cover all these bases with a truncated tree so I
think this justifies the extra storage and load time in the test
database.