Remove incorrect energy_correction from pdep.py
After finding many submerged barriers when running RMG in pdep mode, I discovered that the energy_correction is calculated incorrectly *and* only applied to the transition state. The code calculates the energy_correction by averaging all species E0s. However, it should average the total energy for each stationary point (some of which are bimolecular), and I believe it would be more intuitive to use the minimum energy on the surface. A simple solution would be to remove the energy_correction altogether (what I've done here for now). If it is desired, we need to apply energy_correction to all stationary points, not just transition states, so that barriers are not submerged. 

The energy_correction could be calculated as follows...
```
stationary_point_E0s = [sum([spec.conformer.E0.value_si for spec in stationary_point]) for stationary_point  in self.reactants + self.isomers + self.products])]
energy_correction = -np.array(stationary_point_E0s).min()
```

I would need to track down where to apply it to the wells though.