Let Regression tests run concurrently with Unit tests.
Just to save wall clock time, we could let the regression tests
start before the full matrix of unit tests is complete.
Some of them (especially the mac os ones) take a couple of hours.

Possible drawback is we'll get comment annotations about the regression 
tests even if the unit tests fail. But maybe that's not a bad thing.

The build and push docker now requires both build-and-test as well 
as regression-test to succeed.