Set PYTHON_JULIAPKG_PROJECT for Docker
Being in a Conda environment is effectively equivalent to setting:
   export PYTHON_JULIAPKG_PROJECT="$CONDA_PREFIX/julia_env"
The JuliaPy PythonJulia stack (and juliapkg in particular) tries to automatically detect whether it's running inside a Conda or Python virtual environment. 
But when building a Docker image we're not _exactly_ in a conda environment
we're kind of faking it, and so it doesn't detect it and put stuff
in the right place. So in the docker context we have to set it explicitly.
Let's make it explicit everywhere.