Only run docker build on PRs from official fork.
Because the Dockerfile gets the RMG-Py source via

  git clone --single-branch --branch ${RMG_Py_Branch} --depth 1 https://github.com/ReactionMechanismGenerator/RMG-Py.git && \

it only works if the ${RMG_Py_Branch} is on the official fork at
https://github.com/ReactionMechanismGenerator/RMG-Py.git
and will not work on a PR coming from a fork.
Thus, we should only try running it on PRs coming from the official fork.
Hopefully that's what this tweak does.