Use lowest energy Configuration for energy_correction.
As noted in https://github.com/ReactionMechanismGenerator/RMG-Py/pull/2791
the energy_correction was not actually an average of the Configuration energies
because some Configurations (reactants and products) are bimolecular.
This corrects the calculation, but also switches to using the MINIMUM
rather than the average, so that the lowest should end up at zero.