Remove energy_correction from TS E0 values when writing Arkane input file
The E0 values for TS have been modified using the energy_correction, but the
E0 values for species have not been. We can't change the species, because it needs
to be applied to the Configuration (some are bimolecular, some are unimolecular),
so to be consistent we must remove it from the TS.
This leaves both values, and a comment.
It's a bit ugly, and I think there should be a better fix later,
but this I think solves an immediate bug, and now the network_X_Y.py files 
written during an RMG job are at least internally consistent.
(They have a different definition of 0 from inside the RMG job, but they're self-consistent.)