Make drawing all PES diagrams an option in the input file.
You can set generatePESDiagrams=True in the options() block
of an input file if you want to save them all.
This is presumably a runtime penalty, or at least a hard disk space penalty,
and probably shouldn't be a default.
Default is False.

Fixup:
When the pdep_settings was made from an Arkane explorer job rather
than an RMG run, the generate_PES_diagrams attribute does not exist.
So we now assume False if not specified.